requirements.txt
requests2==2.16.0
pandas==1.0.1
beautifulsoup4==4.8.2